Add  support for SPA fallbacks
The core issue is with client-side routing. In a Single-Page Application (SPA), the JavaScript framework handles the navigation. When you click a link to go from `/` to `/about`, the browser's URL changes but the framework intercepts it and just updates the page contents without making a new request.

However, if you are on the `/about` page and press the refresh button, the browser sends a `GET /about` request, which will fail as our server looks for a non-existing `/about.html` and will return a 404.

The solution is to fallback to sending a `index.html` instead of 404 for single-page apps. For regular webpages the regular behaviour should suffice